:recycle: 重构代码 - 重构 `handle_date` 相关逻辑
