+++ src/plugins/ELF_RSS2/RSS/routes/Parsing/check_update.py
@@ -1,23 +1,9 @@
 import hashlib
-import time
 
 from tinydb import TinyDB, Query
 from typing import Dict, Any
 
-
-# 处理日期
-async def handle_date(date=None) -> str:
-    if date:
-        if not isinstance(date, tuple):
-            date = tuple(date)
-        rss_time = time.mktime(date)
-        # 时差处理，待改进
-        if rss_time + 28800.0 < time.time():
-            rss_time += 28800.0
-        return "日期：" + time.strftime("%m月%d日 %H:%M:%S", time.localtime(rss_time))
-    # 没有日期的情况，以当前时间
-    else:
-        return "日期：" + time.strftime("%m月%d日 %H:%M:%S", time.localtime())
+from .handle_date import handle_date
 
 
 # 对 dict 对象计算哈希值，供后续比较
@@ -48,12 +34,10 @@ async def check_update(db: TinyDB, new: list) -> list:
 
     # 对结果按照发布时间排序
     result_with_date = [
-        (await handle_date(i.get("updated_parsed")), i)
-        if i.get("updated_parsed")
-        else (await handle_date(i.get("published_parsed")), i)
+        (await handle_date(i.get("published_parsed", i.get("published_parsed"))), i)
         for i in to_send_list
     ]
-    result_with_date.sort(key=lambda tup: tup[0])
+    result_with_date.sort()
     result = [i for key, i in result_with_date]
 
     return result
